requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.115.12
frozenlist==1.5.0
h11==0.16.0
idna==3.10
jinja2==3.1.6
MarkupSafe==2.1.5
multidict==6.1.0
outcome==1.3.0.post0
propcache==0.2.0
pydantic==2.10.6
pydantic-core==2.27.2
PySocks==1.7.1
python-decouple==3.8
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.44.0
trio==0.27.0
trio-websocket==0.12.2
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.15.2
